Moved order related stuff from workers.go to order.go. Created type for list of orders.
