+++ simulator/orders.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"cloud.google.com/go/bigquery"
+	"google.golang.org/api/iterator"
 )
 
 // Type for creating an order, adding products to it, 
@@ -21,6 +22,7 @@ type Order struct {
 	delivery_type int
 	status        int
 }
+type Orders []Order
 
 // Type for temporarily storing order-data returned by the database.
 type OrderReceiver struct {
@@ -182,4 +184,98 @@ func (order *Order) Send(ctx context.Context, client *bigquery.Client) error {
 	}
 
 	return nil
+}
+
+func GetOpenOrders(ctx context.Context, client *bigquery.Client) (Orders, error) {
+    // TODO: Move ids to config file somewhere.
+    project_id := "nettikauppasimulaattori"
+    dataset_id := "store_operational"
+    table_id := "orders"
+
+    sql := fmt.Sprintf("SELECT id, customer_id, delivery_type, status, order_placed FROM `%s.%s.%s` WHERE status = %d",
+        project_id, dataset_id, table_id, ORDER_PENDING)
+    slog.Debug(sql)
+
+    var orders Orders
+
+    // slog.Debug("Sending query.")
+    q := client.Query(sql)
+    job, err := q.Run(ctx)
+    if err != nil { 
+        slog.Error(fmt.Sprint(err))
+        return orders, err }
+
+    // slog.Debug("Wait query.")
+    status, err := job.Wait(ctx)
+    if err != nil { 
+        slog.Error(fmt.Sprint(err))
+        return orders, err 
+    }
+
+    // slog.Debug("Check status.")
+    if status.Err() != nil { 
+        slog.Error(fmt.Sprint(err))
+        return orders, status.Err()
+    }
+
+    // slog.Debug("Get iterator.")
+    it, err := job.Read(ctx)
+    if err != nil { 
+        slog.Error(fmt.Sprint(err))
+        return orders, err
+    }
+
+    // slog.Debug("Parse results.")
+    for {
+        var order OrderReceiver
+        if it.Next(&order) == iterator.Done { break }
+        // fmt.Printf("%d: %T\n", tmp.ID, tmp.ID)
+        orders = append(orders, ConvertOrderReceiverToOrder(order))
+    }
+
+    // TODO: Add error if res has zero length.
+
+    return orders, nil
+}
+
+func ConvertOrderReceiverToOrder(o OrderReceiver) Order {
+    var res Order
+
+    res.id = o.id
+    res.customer_id = o.customer_id
+    res.order_placed = o.order_placed
+    res.delivery_type = o.delivery_type
+    res.status = o.status
+        
+    return res
+}
+
+
+func UpdateOrder(order Order, ctx context.Context, client *bigquery.Client) error {
+    project_id := "nettikauppasimulaattori"
+    dataset_id := "store_operational"
+    table_id := "orders"
+
+    now, _ := nowInTimezone("Europe/Helsinki")
+
+    sql := fmt.Sprintf("UPDATE `%s.%s.%s` SET status = %d, shipping_date = \"%s\", last_modified = \"%s\", tracking_number = %d WHERE id = %d",
+        project_id, 
+        dataset_id,
+        table_id, 
+        ORDER_SHIPPED,
+        Time2SQLDate(now), 
+        Time2SQLDatetime(now), 
+        rand.Int(),
+        order.id)
+    slog.Debug(sql)
+
+    q := client.Query(sql)
+    job, err := q.Run(ctx)
+    if err != nil { return err }
+
+    status, err := job.Wait(ctx)
+    if err != nil { return err }
+    if status.Err() != nil { return status.Err() }
+
+    return nil
 }
